backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-redis
Pillow
redis